seperated the embed helper functions and refactored the references to them
