+++ commands/economy/leaderboard.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require("discord.js");
-const { embedReplyPrimaryColor, embedReplyFailureColor } = require("../../helpers/embed-reply");
+const { embedReplyPrimaryColor, embedReplyFailureColor } = require("../../helpers/embeds/embed-reply");
 const db = require("../../helpers/db");
 const { logToFileAndDatabase } = require("../../helpers/logger");
