+++ commands/economy/pay.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require("discord.js");
-const { embedReplyFailureColor, embedReplySuccessColor } = require("../../helpers/embed-reply");
+const { embedReplyFailureColor, embedReplySuccessColor } = require("../../helpers/embeds/embed-reply");
 const { logToFileAndDatabase } = require("../../helpers/logger");
 const db = require("../../helpers/db");
