+++ commands/utility/say.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require("discord.js");
-const { embedReplySaidByPrimaryColor } = require("../../helpers/embed-reply");
+const { embedReplySaidByPrimaryColor } = require("../../helpers/embeds/embed-reply");
 const { logToFileAndDatabase } = require("../../helpers/logger");
 
 module.exports = {
